+++ src/sub_commands.rs
@@ -3,3 +3,4 @@ pub mod c_build;
 pub mod c_varbuild;
 pub mod c_migrate;
 pub mod c_dev;
+pub mod c_version;